% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeMoments}
\alias{computeMoments}
\title{computeMoments}
\usage{
computeMoments(X, Y)
}
\arguments{
\item{X}{Matrix of input data (size nxd)}

\item{Y}{Vector of binary outputs (size n)}
}
\value{
A list L where L[[i]] is the i-th cross-moment
}
\description{
Compute cross-moments of order 1,2,3 from X,Y
}
\examples{
X <- matrix(rnorm(100), ncol=2)
Y <- rbinom(100, 1, .5)
M <- computeMoments(X, Y)

}
