% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(X)
}
\arguments{
\item{X}{Vector or matrix to be normalized}
}
\value{
The normalized matrix (1 column if X is a vector)
}
\description{
Normalize a vector or a matrix (by columns), using euclidian norm
}

