% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotBox}
\alias{plotBox}
\title{plotBox}
\usage{
plotBox(mr, x, y, ...)
}
\arguments{
\item{mr}{Output of multiRun(), list of lists of functions results}

\item{x}{Row index of the element inside the aggregated parameter}

\item{y}{Column index of the element inside the aggregated parameter}

\item{...}{Additional graphical parameters (xlab, ylab, ...)}
}
\description{
Draw compared boxplots of a single parameter (scalar)
}
\examples{
\dontrun{
beta <- matrix(c(1,-2,3,1),ncol=2)
mr <- multiRun(...) #see bootstrap example in ?multiRun
                    #mr[[i]] is a list of estimated parameters matrices
mu <- normalize(beta)
for (i in 1:2)
  mr[[i]] <- alignMatrices(res[[i]], ref=mu, ls_mode="exact")
plotBox(mr, 2, 1) #second row, first column}

}
