% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotQn}
\alias{plotQn}
\title{plotQn}
\usage{
plotQn(N, n, p, beta, b, link)
}
\arguments{
\item{N}{Number of starting points}

\item{n}{Number of points in sample}

\item{p}{Vector of proportions}

\item{beta}{Regression matrix (target)}

\item{b}{Vector of biases}

\item{link}{Link function (logit or probit)}
}
\description{
Draw 3D map of objective function values
}

