% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alignMatrices}
\alias{alignMatrices}
\title{alignMatrices}
\usage{
alignMatrices(Ms, ref, ls_mode)
}
\arguments{
\item{Ms}{A list of matrices, all of same size DxK}

\item{ref}{Either a reference matrix or "mean" to align on empirical mean}

\item{ls_mode}{How to compute the labels assignment: "exact" for exact algorithm
(default, but might be time-consuming, complexity is O(K^3) ), or "approx1", or
"approx2" to apply a greedy matching algorithm (heuristic) which for each column in
reference (resp. in current row) compare to all unassigned columns in current row
(resp. in reference)}
}
\value{
The aligned list (of matrices), of same size as Ms
}
\description{
Align a set of parameters matrices, with potential permutations.
}

