% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotBox}
\alias{plotBox}
\title{plotBox}
\usage{
plotBox(mr, x, y)
}
\arguments{
\item{mr}{Output of multiRun(), list of lists of functions results}

\item{x}{Row index of the element inside the aggregated parameter}

\item{y}{Column index of the element inside the aggregated parameter}
}
\description{
Draw boxplot
}
\examples{
#See example in ?plotHist
}

