% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCoefs}
\alias{plotCoefs}
\title{plotCoefs}
\usage{
plotCoefs(mr, params)
}
\arguments{
\item{mr}{Output of multiRun(), list of lists of functions results}

\item{params}{True value of parameters matrix}
}
\description{
Draw coefs estimations + standard deviations
}
\examples{
#See example in ?plotHist
}

